Fix appearance of percent bars
Before this commit, the percent bars had a few pixels padding away from
the bottom of the row for torrents that were not displaying a message,
and no padding for torrents that were displaying a message.  Now there
will always be 4px in between the percent bar and the bottom of the row.