Prevent mouse wheel from scrolling main page when a dialog is showing
This works by:
 - preventing mouse wheel events generated by the main page
 - notifying the main page of mouse wheel events generated by child
   frames, and canceling the scroll event on the main window

I'm not very satisfied with this approach.  It seems like handling
scroll events properly on a child iframe is a tricky case.  What I've
done seems to work pretty well (naturally, except for in IE).  However,
my solution has the problem that the scroll events sent to the main page
(which are canceled) cause the child frame to lose the focus so further
scroll events go directly to the main page until the user clicks on the
child frame again.